refactor : 리뷰해주신 내용#5 - SubwayMap 객체 사용
