+++ src/main/java/nextstep/subway/line/domain/PathFinder.java
@@ -4,15 +4,12 @@ import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.exception.EqualStationException;
 import nextstep.subway.station.exception.NoStationException;
 import org.jgrapht.GraphPath;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class PathFinder {
 
@@ -25,15 +22,8 @@ public class PathFinder {
     public StationGraph getPathInfo(Station source, Station target) {
         validateNullStation(source, target);
         validateEqualStation(source, target);
-
-        WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
-
-        for (Line line : lines) {
-            setGraph(graph, line);
-        }
-
-        DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-        GraphPath pathInfo = dijkstraShortestPath.getPath(source, target);
+        SubwayMap subWayMap = new SubwayMap(new WeightedMultigraph(DefaultWeightedEdge.class));
+        GraphPath pathInfo = subWayMap.getPath(lines, source, target);
         return new StationGraph(pathInfo);
     }
 
@@ -47,36 +37,5 @@ public class PathFinder {
         if (source == target) {
             throw new EqualStationException();
         }
-    }
-
-    private void setGraph(WeightedMultigraph<Station, DefaultWeightedEdge> graph, Line line) {
-        List<Section> sections = getSections(line);
-        setVertex(graph, sections);
-        setEdgeWeight(graph, sections);
-    }
-
-    private List<Section> getSections(Line line) {
-        return line.getSections().getSections();
-    }
-
-    private void setVertex(WeightedMultigraph<Station, DefaultWeightedEdge> graph, List<Section> sections) {
-        List<Station> stations = getStations(sections);
-
-        for (Station station : stations) {
-            graph.addVertex(station);
-        }
-    }
-
-    private void setEdgeWeight(WeightedMultigraph<Station, DefaultWeightedEdge> graph, List<Section> sections) {
-        for (Section section : sections) {
-            graph.setEdgeWeight(graph.addEdge(section.getUpStation(), section.getDownStation()), section.getDistance());
-        }
-    }
-
-    private List<Station> getStations(List<Section> sections) {
-        return sections.stream()
-                .flatMap(s -> Stream.of(s.getUpStation(), s.getDownStation()))
-                .distinct()
-                .collect(Collectors.toList());
     }
 }
